Workspace - fix few issues, more to come.
New ProjectWorkspace which holds list of projects. Works now for
uvision, that it builds 2 projects to own folder. I'll fix it that
it will generate two projects to the workspace directory and create aborts
workspace. We need to redesign how Project, Workspace are implemented,
it's quite messy , to edit or add new features like this