Project - 2 dictionaries - common and tool_specific
This reflects yaml structure. Although the common has a mixture of
internal data at the moment. Once we do export, we use common,
which is at the time of the export a merge of common and tool specific
dictionaries, this might change and we decouple this data to:

- common
- tool_specific
- exported (merged common and tool)

We can print in debug all 3 so if there's a problem, user can examine all 3 data.