Project workspace - rework how they work
Workspace is on its own (workspaces), as they are different. Most used
feature are projects, which is unchanged, I'll keep working on workspaces,
to enable them for cmake for instance. This commit fixes uvision workspaces.

in projects.yaml:

```
workspaces:
    workspace_test:
        projects:
            - k20d50m_blinky
            - lpc1768_blinky
        settings:
            export_dir:
                - generated_projects\workspaces\{workspace_name}
```
This defines one workspace, 2 projects and export dir for workspace (overwrites projects export_dirs).