IAR - ordereddict for ewp/ewd files, indent 2 spaces, encoding as IAR
IAR has own ordered xml, thus we should follow it, and then IAR when you open a project,
does not change the entire file. We now generate as close as possible to the xml IAR format.
This fixes debugger settings as I noticed IAR ignored them (ewd file require a name in the config!).