Move validation code into the argument parser
The arguments that were taken from a list of keys in upper or lower case are now converted before lookup.
The arguments that were passed to os.path.exists() are now validated in the argument parser.
When an argument error occurs, the return code is no longer -1, but the more standard 2.

This allows us to provide more descritpive and accurate error messages to the user,
and assume that if arguments passed validation they exist in the filesystem or will
always work as indexes into the dict.

before :
```
$ progen generate -t uvision -f foo
$ echo $?
255
```

after :
```
$ progen generate -t uvision -f foo
usage: progen generate [-h] [-f FILE] [-p PROJECT]
                       [-t {visual_studio_make_gcc_arm,gdb,uvision5,uvision4,sublime_make_gcc_arm,jlink_gdb,iar_arm,coide,make_gcc_arm,eclipse_make_gcc_arm,visual_studio_gdb,arm_none_eabi_gdb,cmake_gcc_arm,make_armcc,visual_studio,cmake,sublime,sublime_text,make_gcc,eclipse,gcc_arm,uvision,iar}]
                       [-b] [-c]
progen generate: error: argument -f/--file: foo is not a file.
$ echo $?
2
```